Fix error in PrintRegion::flow() port

+++ xs/src/libslic3r/PrintRegion.cpp
@@ -60,7 +60,7 @@ PrintRegion::flow(FlowRole role, double layer_height, bool bridge, bool first_la
     }
     double nozzle_diameter = this->_print->config.nozzle_diameter.get_at(extruder-1);
     
-    return Flow::new_from_config_width(role, config_width, nozzle_diameter, layer_height, bridge ? this->config.bridge_flow_ratio : 0);
+    return Flow::new_from_config_width(role, config_width, nozzle_diameter, layer_height, bridge ? (float)this->config.bridge_flow_ratio : 0.0);
 }
 
 #ifdef SLIC3RXS
